Improve `executable` module structure
